requirements.txt
absl-py==2.0.0
altair==5.1.2
astunparse==1.6.3
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
flatbuffers==20181003210633
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.14.0
libclang==16.0.6
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.2.0
numpy==1.26.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.1
Pillow==10.1.0
plotly==5.18.0
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.27.2
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-cpu-aws==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
